pyproject.toml
[project]
name = "stage_1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.1",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "resend>=2.21.0",
    "sqlalchemy>=2.0.46",
    "uuid7>=0.1.0",
]
